+++ lib/src/pods/shared_pods/shared_string_list_pod.dart
@@ -49,4 +49,5 @@ final class SharedStringListPodCreator {
 
 typedef TSharedStringListPod = SharedPod<List<String>, List<String>>;
 typedef TSharedTempStringListPod = SharedTempPod<List<String>, List<String>>;
-typedef TSharedGlobalStringListPod = SharedGlobalPod<List<String>, List<String>>;
+typedef TSharedGlobalStringListPod
+    = SharedGlobalPod<List<String>, List<String>>;