+++ tests/test_basics/lib/pages/future_counter_test.dart
@@ -13,7 +13,8 @@ import 'package:flutter/material.dart';
 
 // ░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░
 
-final _pFutureCounter = Future.delayed(const Duration(seconds: 3), () => GlobalPod<int>(1));
+final _pFutureCounter =
+    Future.delayed(const Duration(seconds: 3), () => GlobalPod<int>(1));
 
 class FutureCounterTest extends StatelessWidget {
   const FutureCounterTest({super.key});